requirements.txt
numpy==2.2.0
pandas==2.2.3
pillow==11.0.0